fix subscription tests move GCValue stuff out of the connector
