+++ server/connectors/api-connector/src/main/scala/com/prisma/api/connector/NodeSelector.scala
@@ -1,8 +1,7 @@
 package com.prisma.api.connector
 
-import com.prisma.gc_values.{GCValue, IdGCValue}
+import com.prisma.gc_values.{GCValue, GCValueExtractor, IdGCValue}
 import com.prisma.shared.models.{Field, Model}
-import com.prisma.util.gc_value.{GCAnyConverter, GCValueExtractor}
 
 object NodeSelector {
   def forId(model: Model, id: String): NodeSelector                  = NodeSelector(model, model.getFieldByName_!("id"), IdGCValue(id))
@@ -12,17 +11,6 @@ object NodeSelector {
 
 case class NodeSelector(model: Model, field: Field, fieldValue: GCValue) {
   lazy val fieldName                  = field.name
-  lazy val unwrappedFieldValue: Any   = GCValueExtractor.fromGCValue(fieldValue)
   lazy val fieldValueAsString: String = GCValueExtractor.fromGCValueToString(fieldValue)
   lazy val isId: Boolean              = field.name == "id"
-
-  def updateValue(value: Any): NodeSelector = {
-    val unwrapped = value match {
-      case Some(x) => x
-      case x       => x
-    }
-
-    val newGCValue = GCAnyConverter(field.typeIdentifier, isList = false).toGCValue(unwrapped).get
-    this.copy(fieldValue = newGCValue)
-  }
 }
