+++ server/servers/api/src/main/scala/com/prisma/api/resolver/ScalarListDeferredResolver.scala
@@ -2,10 +2,9 @@ package com.prisma.api.resolver
 
 import com.prisma.api.connector.{DataResolver, ScalarListValues}
 import com.prisma.api.resolver.DeferredTypes._
-import com.prisma.gc_values.IdGCValue
-import com.prisma.util.gc_value.GCValueExtractor
-
+import com.prisma.gc_values.GCValueExtractor
 import scala.concurrent.ExecutionContext.Implicits.global
+
 import scala.concurrent.Future
 
 class ScalarListDeferredResolver(dataResolver: DataResolver) {
