+++ server/servers/api/src/main/scala/com/prisma/subscriptions/resolving/FilteredResolver.scala
@@ -3,6 +3,7 @@ package com.prisma.subscriptions.resolving
 import com.prisma.api.connector.Types.DataItemFilterCollection
 import com.prisma.api.connector._
 import com.prisma.api.schema.ObjectTypeBuilder
+import com.prisma.gc_values.IdGCValue
 import com.prisma.shared.models.Model
 import sangria.schema.Context
 
@@ -29,7 +30,8 @@ object FilteredResolver {
         case _                => true
       }
 
-    val filter = filterInput.filter(removeTopLevelIdFilter(_)) ++ List(FilterElement(key = "id", value = id, field = Some(model.getFieldByName_!("id"))))
+    val filter = filterInput.filter(removeTopLevelIdFilter(_)) ++ List(
+      FinalValueFilter(key = "id", value = IdGCValue(id), field = model.getFieldByName_!("id"), ""))
 
     dataResolver.resolveByModel(model, Some(QueryArguments.filterOnly(filter = Some(filter)))).map(_.nodes.headOption)
   }
